[test] stdlib/StringIndex: Spin off O(n^4)  substring replacement test into a standalone long test
Also trim down its input a bit so that this doesn’t take 20 minutes.
